+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+### Added
+* Added `model_type` option to `translate()` to use models with higher
+  translation quality (available for some language pairs), or better latency.
+  Options are `'quality_optimized'`, `'latency_optimized'`, and  `'prefer_quality_optimized'`
+* Added the `model_type_used` field to the `translate()` response, that
+  indicates the translation model used when the `model_type` option is
+  specified.
+
+
 ## [3.0.2] - 2024-10-02
 ### Added
 * Added doc example and tests for context parameter